// ford-2.0-app-dev/ford-2.0-app-dev/src/model/Configuration.ts
export interface Configuration {
    testDriveDemarcationOwnerUrl?: string;
    testDriveDemarcationFordUrl?: string;
    testDriveDemarcationOwnerInCaravanUrl?: string;
    testDriveTermsUrl?: string;
    newsletterUrl?: string;
    quoteUrl?: string;
    contactData?: string;
}

export interface ConfigurationDB {
    testDriveDemarcationOwnerUrl?: string;
    testDriveDemarcationFordUrl?: string;
    testDriveDemarcationOwnerInCaravanUrl?: string;
    testDriveTermsUrl?: string;
    newsletterUrl?: string;
    quoteUrl?: string;
    contactData?: string;
}